requirements.txt
python-telegram-bot[job-queue]==20.3
aiohttp>=3.8.0
